CRM-13244 - RecipientBuilder - Move entity-specific bits to ActionMappings
